requirements.txt
moviepy>=1.0.3,<2.0.0
Pillow>=10.0.0
instagrapi>=2.0.0
python-dotenv>=1.0.0
opencv-python-headless>=4.0.0
